Version 1.8

+++ app/src/main/java/com/multimedia/aes/gestnet_nucleo/dao/ManoObraDAO.java
@@ -22,7 +22,7 @@ public class ManoObraDAO extends DBHelperMOS {
 
     //__________FUNCIONES DE CREACIÓN________________________//
 
-    public static boolean newManoObra(Context context,int id_mano, String concepto, int precio, String coste )throws java.sql.SQLException {
+    public static boolean newManoObra(Context context,int id_mano, String concepto, double precio, String coste )throws java.sql.SQLException {
         ManoObra d = montarManoObra( id_mano, concepto,precio,coste);
         return crearManoObra(d,context);
     }
@@ -36,7 +36,7 @@ public class ManoObraDAO extends DBHelperMOS {
             return false;
         }
     }
-    public static ManoObra montarManoObra(   int id_mano, String concepto, int precio, String coste ) {
+    public static ManoObra montarManoObra(   int id_mano, String concepto, double precio, String coste ) {
         ManoObra d =new ManoObra(id_mano, concepto,precio,coste);
         return d;
     }
@@ -99,7 +99,7 @@ public class ManoObraDAO extends DBHelperMOS {
             return listadoManoObra.get(0);
         }
     }
-    public static int buscarPrecioManoObraPorNombre(Context context, String s) throws java.sql.SQLException {
+    public static double buscarPrecioManoObraPorNombre(Context context, String s) throws java.sql.SQLException {
         cargarDao(context);
         List<ManoObra> listadoManoObra= null;
         try {
@@ -116,7 +116,7 @@ public class ManoObraDAO extends DBHelperMOS {
     }
 
     //____________________________FUNCIONES DE ACTUALIZAR_________________________________________//
-    public static void actualizarManoObra(Context context, int id_mano, String concepto, int precio, String coste ) throws java.sql.SQLException {
+    public static void actualizarManoObra(Context context, int id_mano, String concepto, double precio, String coste ) throws java.sql.SQLException {
         cargarDao(context);
         UpdateBuilder<ManoObra, Integer> updateBuilder = dao.updateBuilder();
         updateBuilder.where().eq(ManoObra.ID_MANO,id_mano);
